Refactor de la arquitectura del proyecto
